+++ lib/generators/simple_form/templates/config/initializers/simple_form.rb
@@ -143,6 +143,9 @@ SimpleForm.setup do |config|
   # Default class for inputs
   # config.input_class = nil
 
+  # Define the default class of the input wrapper of the boolean input.
+  config.boolean_label_class = 'checkbox'
+
   # Defines if the default input wrapper class should be included in radio
   # collection wrappers.
   # config.include_default_input_wrapper_class = true