+++ test/inputs/boolean_input_test.rb
@@ -128,12 +128,32 @@ class BooleanInputTest < ActionView::TestCase
     end
   end
 
+  test 'input boolean with nested style works using :input only in wrapper config (no label_input), adding the extra label wrapper with custom class' do
+    swap_wrapper do
+      swap SimpleForm, :boolean_style => :nested, :boolean_label_class => 'foo' do
+        with_input_for @user, :active, :boolean
+
+        assert_select 'label.boolean + input[type=hidden] + label.foo > input.boolean'
+      end
+    end
+  end
+
   test 'input boolean with nested style works using :label_input in wrapper config, adding "checkbox" class to label' do
     swap_wrapper :default, self.custom_wrapper_without_top_level do
       swap SimpleForm, :boolean_style => :nested do
         with_input_for @user, :active, :boolean
 
         assert_select 'input[type=hidden] + label.boolean.checkbox > input.boolean'
+      end
+    end
+  end
+
+  test 'input boolean with nested style works using :label_input in wrapper config, adding custom class to label' do
+    swap_wrapper :default, self.custom_wrapper_without_top_level do
+      swap SimpleForm, :boolean_style => :nested, :boolean_label_class => 'foo' do
+        with_input_for @user, :active, :boolean
+
+        assert_select 'input[type=hidden] + label.boolean.foo > input.boolean'
       end
     end
   end